Добавила в константы дефолтное значение фильтра и клавиши esc, перенесла FilterValues в модуль карты
